LPS-54556 Move RSSDisplayContext and remove old configuration
