+++ modules/apps/rss/rss-web/src/META-INF/resources/feed.jspf
@@ -22,19 +22,19 @@
 			SyndImage syndImage = syndFeed.getImage();
 			%>
 
-			<c:if test="<%= (syndImage != null) && rssDisplayContext.isShowFeedImage() %>">
+			<c:if test="<%= (syndImage != null) && rssPortletInstanceSettings.showFeedImage() %>">
 				<div class="feed-image">
 					<img alt="<%= HtmlUtil.escapeAttribute(syndImage.getDescription()) %>" src="<%= HtmlUtil.escapeHREF(rssFeed.getSyndFeedImageURL()) %>" />
 				</div>
 			</c:if>
 
-			<c:if test="<%= rssDisplayContext.isShowFeedTitle() %>">
+			<c:if test="<%= rssPortletInstanceSettings.showFeedTitle() %>">
 				<div class="feed-title">
 					<aui:a href="<%= HtmlUtil.escapeJSLink(rssFeed.getSyndFeedLink()) %>" target="_blank"><%= HtmlUtil.escape(rssFeed.getTitle()) %></aui:a>
 				</div>
 			</c:if>
 
-			<c:if test="<%= (syndFeed.getPublishedDate() != null) && rssDisplayContext.isShowFeedDescription() %>">
+			<c:if test="<%= (syndFeed.getPublishedDate() != null) && rssPortletInstanceSettings.showFeedDescription() %>">
 				<div class="feed-date feed-published-date">
 					<liferay-ui:icon
 						iconCssClass="icon-calendar"
@@ -44,7 +44,7 @@
 				</div>
 			</c:if>
 
-			<c:if test="<%= Validator.isNotNull(syndFeed.getDescription()) && rssDisplayContext.isShowFeedDescription() %>">
+			<c:if test="<%= Validator.isNotNull(syndFeed.getDescription()) && rssPortletInstanceSettings.showFeedDescription() %>">
 				<div class="feed-description">
 					<%= HtmlUtil.escape(syndFeed.getDescription()) %>
 				</div>
@@ -58,15 +58,15 @@
 				<liferay-ui:panel-container cssClass="feed-entries" extended="<%= false %>" id="rssFeedsPanelContainer" persistState="<%= false %>">
 
 					<%
-					for (int j = 0; (j < rssFeedEntryDisplayContexts.size()) && (j < rssDisplayContext.getEntriesPerFeed()); j++) {
+					for (int j = 0; (j < rssFeedEntryDisplayContexts.size()) && (j < rssPortletInstanceSettings.entriesPerFeed()); j++) {
 						RSSFeedEntry rssFeedEntry = rssFeedEntryDisplayContexts.get(j);
 
 						SyndEntry syndEntry = rssFeedEntry.getSyndEntry();
 					%>
 
-						<liferay-ui:panel cssClass="feed-entry" defaultState='<%= (themeDisplay.isStateMaximized() || (j < rssDisplayContext.getExpandedEntriesPerFeed())) ? "open" : "closed" %>' extended="<%= false%>" title="<%= HtmlUtil.escape(syndEntry.getTitle()) %>">
+						<liferay-ui:panel cssClass="feed-entry" defaultState='<%= (themeDisplay.isStateMaximized() || (j < rssPortletInstanceSettings.expandedEntriesPerFeed())) ? "open" : "closed" %>' extended="<%= false%>" title="<%= HtmlUtil.escape(syndEntry.getTitle()) %>">
 							<div class="feed-entry-content">
-								<c:if test="<%= rssDisplayContext.isShowFeedItemAuthor() && Validator.isNotNull(syndEntry.getAuthor()) %>">
+								<c:if test="<%= rssPortletInstanceSettings.showFeedItemAuthor() && Validator.isNotNull(syndEntry.getAuthor()) %>">
 									<div class="feed-entry-author">
 										<%= HtmlUtil.escape(syndEntry.getAuthor()) %>
 									</div>
