+++ modules/apps/rss/rss-web/src/META-INF/resources/init.jsp
@@ -28,7 +28,9 @@ page import="com.liferay.portal.kernel.util.FastDateFormatFactoryUtil" %><%@
 page import="com.liferay.portal.kernel.util.HtmlUtil" %><%@
 page import="com.liferay.portal.kernel.util.StringPool" %><%@
 page import="com.liferay.portal.kernel.util.Validator" %><%@
-page import="com.liferay.rss.web.context.RSSDisplayContext" %><%@
+page import="com.liferay.rss.web.display.context.RSSDisplayContext" %><%@
+page import="com.liferay.rss.web.display.context.util.RSSRequestHelper" %><%@
+page import="com.liferay.rss.web.settings.RSSPortletInstanceSettings" %><%@
 page import="com.liferay.rss.web.util.RSSFeed" %><%@
 page import="com.liferay.rss.web.util.RSSFeedEntry" %>
 
@@ -47,7 +49,11 @@ page import="java.util.List" %>
 <portlet:defineObjects />
 
 <%
-RSSDisplayContext rssDisplayContext = new RSSDisplayContext(request, portletPreferences);
+RSSRequestHelper rssRequestHelper = new RSSRequestHelper(request);
+
+RSSPortletInstanceSettings rssPortletInstanceSettings = rssRequestHelper.getRSSPortletInstanceSettings();
+
+RSSDisplayContext rssDisplayContext = new RSSDisplayContext(rssPortletInstanceSettings);
 
 Format dateFormatDateTime = FastDateFormatFactoryUtil.getDateTime(locale, timeZone);
 %>
