+++ modules/apps/rss/rss-web/src/META-INF/resources/view.jsp
@@ -20,7 +20,7 @@
 List<RSSFeed> rssFeeds = rssDisplayContext.getRSSFeeds();
 %>
 
-<liferay-ui:ddm-template-renderer className="<%= RSSFeed.class.getName() %>" displayStyle="<%= rssDisplayContext.getDisplayStyle() %>" displayStyleGroupId="<%= rssDisplayContext.getDisplayStyleGroupId() %>" entries="<%= rssFeeds %>">
+<liferay-ui:ddm-template-renderer className="<%= RSSFeed.class.getName() %>" displayStyle="<%= rssPortletInstanceSettings.displayStyle() %>" displayStyleGroupId="<%= rssPortletInstanceSettings.displayStyleGroupId(themeDisplay.getScopeGroupId()) %>" entries="<%= rssFeeds %>">
 
 	<%
 	for (int i = 0; i < rssFeeds.size(); i++) {
